# Restore performance for the simple case