```
benchmark                     old ns/op     new ns/op     delta
BenchmarkGetBool-4            1021          479           -53.09%
BenchmarkGetBoolFromMap-4     6.56          6.39          -2.59%

benchmark                     old allocs     new allocs     delta
BenchmarkGetBool-4            6              4              -33.33%
BenchmarkGetBoolFromMap-4     0              0              +0.00%

benchmark                     old bytes     new bytes     delta
BenchmarkGetBool-4            113           49            -56.64%
BenchmarkGetBoolFromMap-4     0             0             +0.00%
```

Fixes #249
Fixes https://github.com/spf13/hugo/issues/2536